♻️ 凡例の処理を別ファイルとして切り出した
