+++ packages/shared/src/components/parts/graph.component.tsx
@@ -1,5 +1,6 @@
-import React, { useCallback, useId, useLayoutEffect, useMemo, useRef, useState } from "react";
+import React, { useCallback, useId, useMemo, useRef, useState } from "react";
 import { AggregatedData } from "@fukui-kanko/shared";
+import { ClickableLegend } from "@fukui-kanko/shared/components/parts";
 import {
   ChartContainer,
   ChartLegend,
@@ -7,8 +8,7 @@ import {
   ChartTooltipContent,
 } from "@fukui-kanko/shared/components/ui";
 import { GRAPH_VIEW_TYPES } from "@fukui-kanko/shared/types";
-import { cn, WEEK_DAYS } from "@fukui-kanko/shared/utils";
-import type { LegendProps } from "recharts";
+import { WEEK_DAYS } from "@fukui-kanko/shared/utils";
 import { CartesianGrid, Line, LineChart, XAxis, YAxis } from "recharts";
 
 type GraphProps = {
@@ -27,81 +27,6 @@ type XAxisTickProps = {
 
 const chartConfig = {
   totalCount: { label: "人物検出回数" },
-};
-
-const ClickableLegend: React.FC<{
-  payload?: LegendProps["payload"];
-  hidden: Set<string>;
-  onToggle: (key: string) => void;
-  instanceSuffix: string;
-  savedScrollTop?: number;
-  onScrollPersist?: (top: number) => void;
-  hoveredKey?: string;
-  onHoverKeyChange?: (key?: string) => void;
-}> = ({
-  payload = [],
-  hidden,
-  onToggle,
-  instanceSuffix,
-  savedScrollTop = 0,
-  onScrollPersist,
-  hoveredKey: controlledHoveredKey,
-  onHoverKeyChange,
-}) => {
-  const containerRef = useRef<HTMLDivElement | null>(null);
-  const hoveredKey = controlledHoveredKey;
-
-  const setHover = (key?: string) => {
-    if (onHoverKeyChange) onHoverKeyChange(key);
-  };
-
-  // 再マウント時にスクロール位置を復元（描画前に反映）
-  useLayoutEffect(() => {
-    if (containerRef.current) {
-      containerRef.current.scrollTop = savedScrollTop;
-    }
-  }, [savedScrollTop]);
-
-  const handleScroll = (e: React.UIEvent<HTMLDivElement>) => {
-    onScrollPersist?.(e.currentTarget.scrollTop);
-  };
-  return (
-    <div
-      ref={containerRef}
-      className="flex flex-wrap items-center justify-center gap-4 max-h-[4.5rem] overflow-y-auto"
-      onScroll={handleScroll}
-    >
-      {payload.map((entry) => {
-        const key = `${entry.dataKey}_${instanceSuffix}`;
-        const name = String(entry.value ?? key);
-        const color = entry.color ?? "#999";
-        const isHidden = hidden.has(key);
-        const hoveredIsHidden = hoveredKey ? hidden.has(hoveredKey) : false;
-        const isDimmedByHover = hoveredKey !== undefined && hoveredKey !== key && !hoveredIsHidden;
-        return (
-          <div key={key}>
-            <button
-              onClick={() => {
-                onScrollPersist?.(containerRef.current?.scrollTop ?? 0);
-                onToggle(key);
-              }}
-              onMouseEnter={() => setHover(key)}
-              onMouseLeave={() => setHover(undefined)}
-              onFocus={() => setHover(key)}
-              onBlur={() => setHover(undefined)}
-              className={cn(
-                "flex items-center gap-1.5 cursor-pointer",
-                (isHidden || isDimmedByHover) && "opacity-40",
-              )}
-            >
-              <div className="h-2 w-2 shrink-0 rounded-[2px]" style={{ backgroundColor: color }} />
-              {name}
-            </button>
-          </div>
-        );
-      })}
-    </div>
-  );
 };
 
 function renderTick(props: XAxisTickProps, data: AggregatedData[], xKey: string) {
